Move tests out of basic_test.cc
